fix import statements and moved constants
